fix premium, payment and add docker-compose payment_db

+++ premium-service/src/main/java/com/mss301/premiumservice/model/Plan.java
@@ -39,7 +39,7 @@ public class Plan {
     private int billingCycle;
 
     @Column(name = "price_cents", nullable = false)
-    private int priceCents;
+    private long priceCents;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "currency", nullable = false, length = 10)
